moved socket.io API to the utils folder
