+++ client/src/utils/SocketAPI/index.js
@@ -1,0 +1,3 @@
+import { listenToBoard, sendSquareClickEvent, assignSide } from "./SocketAPI";
+
+export { listenToBoard, sendSquareClickEvent, assignSide };
