+++ routes/userRoutes.js
@@ -3,6 +3,11 @@ const{ User } = require('../models')
 const jwt = require('jsonwebtoken')
 const passport = require('passport')
 
+//get player info
+router.get('/users/:currentJWT',(req,res) => {
+  let myJWT = req.params.currentJWT
+})
+
 // sign in funtionality
 router.post('/users/signUp',(req, res)=>{
   const{ username, leaderboardRanking, userAvatarImageLink } = req.body
@@ -19,4 +24,5 @@ router.post('/users/signUp',(req, res)=>{
      res.json(user ? jwt.sign({id:user._id}, process.env.SECRET) : null)
     })
   })
+
 module.exports = router
